requirements.txt
git+https://github.com/Rapptz/discord.py
bs4
lxml
asyncio
dblpy
psycopg2-binary
requests